[LoggingInterCeptor]: Move logging.interceptor file to interceptors folder
