+++ src/logger/logger.module.ts
@@ -1,7 +1,7 @@
 import { Global, Module } from '@nestjs/common';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 import { CustomLoggerService } from './logger.service';
-import { LoggingInterceptor } from './logging.interceptor';
+import { LoggingInterceptor } from '../interceptors/logging.interceptor';
 
 @Global()
 @Module({
